完成

+++ src/components/TrashDetail.vue
@@ -53,6 +53,12 @@ export default {
     this.getTrashNotes()
       .then(() => {
         this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
+        this.$router.replace({
+          path:'/trash',
+          query:{
+            noteId:this.curTrashNote.id
+          }
+        })
       })
     
   },
@@ -79,9 +85,34 @@ export default {
     
     onRevert() {
       this.revertTrashNote({ noteId: this.curTrashNote.id })
+      .then(()=>{
+         this.setCurTrashNote()
+        this.$router.replace({
+          path:'/trash',
+          query:{
+            noteId:this.curTrashNote.id
+          }
+        })
+      })
+      
     },
     onDelete() {
-      this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        this.$confirm("删除后将无法恢复", "确定删除", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      })
+        .then(() => {
+          return this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        }).then(()=>{
+         this.setCurTrashNote()
+        this.$router.replace({
+          path:'/trash',
+          query:{
+            noteId:this.curTrashNote.id
+          }
+        })
+      })
     },
   },
    beforeRouteUpdate(to, from, next) {
